Remove duplicate get_monomorphized test functions
